[processor/k8sattributes] Move generic testing logic to separate package
The k8sattributes processor has an e2e test with logic that is not specific
to the processor itself. This testing logic will be helpful for other k8s
e2e tests, so this moves that logic into its own package.
