+++ .github/dependabot.yml
@@ -463,6 +463,11 @@ updates:
       interval: "weekly"
       day: "wednesday"
   - package-ecosystem: "gomod"
+    directory: "/internal/k8stest"
+    schedule:
+      interval: "weekly"
+      day: "wednesday"
+  - package-ecosystem: "gomod"
     directory: "/internal/kubelet"
     schedule:
       interval: "weekly"
